+++ src/Metadata/OpenLibrary/OpenLibraryCache.php
@@ -215,7 +215,10 @@ class OpenLibraryCache extends BaseCache
             default => throw new Exception('Invalid cache type'),
         };
         if ($this->hasCache($cacheFile)) {
-            return $this->loadCache($cacheFile);
+            $entry = $this->loadCache($cacheFile);
+            return match ($cacheType) {
+                default => $entry,
+            };
         }
         return null;
     }